refactor: rewrite AspectRatio to use styles
